refactor(auth): resolve scoped abilities from code matrix, not Bouncer
Scoped (publication/submission) role -> ability resolution now reads a
code-owned matrix (App\Auth\RoleAbilities) directly, instead of seeding the
map into Bouncer and reading it back via getAbilities(). The matrix is the
single source of truth: scoped permission changes ship in code and are live
on deploy, with no seeding, convergence, or drift.

- Add App\Auth\RoleAbilities holding the scoped role -> ability MATRIX.
- AbilityResolver reads RoleAbilities directly; drop the Bouncer round-trip,
  the per-instance ability cache, and the BouncerModels dependency.
- AbacSeeder no longer seeds scoped abilities; it keeps only the role rows
  (for GraphQL Role.name) and the global application-administrator wildcard.
- Update authorization + design docs to describe the code-matrix resolution.

Bouncer is retained for what it is good at: the global app-admin grant, role
rows, and (future) runtime-editable global abilities. Scoped resolution is
pure code. Full backend suite green (583).
